oe-layersetup: configs: add meta-clang into the kirkstone configs
Add meta-clang into the 3 kirkstone configs.  meta-clang is pinned to
the kirkstone-clang12 branch because this is what we will be using for
the chromium v111 browser build.

We want to pull meta-clang into all kirkstone builds even without
chromium so that we find any integration problems that clang might
introduce even before we start doing the much larger chromium builds.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>